Allow using $ for inline (html) math with Mathjax
This makes the html and tex versions of the formulas exactly the same,
so we can now have a single macro take some TeX formula that can be
used for both html and tex.

texi2html is updated to allow using $ for the inline
delimiters. extensions.texi updated with a macro to allow writing math
formulas just once that works for both html and tex.