Modify how highlighting of navbar item is done
Each list item of the navbar has an id of "nav-name".  And each page
must have a main tag with id "content-name".  This main tag id is used
to find the name which is then used to find the nav item.  This is
highlighted.

Add comments about this in the file.