Modify CSS to make it work like I want.
o Make the button column narrower so that it's just gray for the width
  of the hamburger menu.
o The navbar isn't fixed position because we want it to scroll.  When
  it was fixed, we couldn't scroll the long list and see all of it on
  small screens.
o Adjust width of various things so that it's wide enough.  90% or
  100% wasn't wide enough.
o Set transform to a larger value so that more of the navbar is
  off-screen. Previously, we see some of the gray on the left, which
  is annoying.