Describing barnch cuts in Cmucl
This is first cut at describing branch cuts in cmucl.  To do that
nicely, we want to be able to display nice equations in pdf and html.

For html, we use MathJax.  This requires some setup.  Therefore, we
tell makeinfo to use texi2html.init file to set up things to use
mathjax.  To make editing equations a little easier, we need to
preprocess .m4 files to create the .texi file.  An m4 macro,
"m4_mathjax" is used to handle formulas for tex, html, and info.

The Makefile is set up to convert texti.m4 to texi files.