Fix #40: FAQ is very wide on mobile
The wide page is caused by the long lines in the `<pre>` blocks.  We
fix this by changing all the `<pre>` blocks to have class "code", if
they weren't already.  Then adjust the CSS for the "code" class to add
a horizontal scrollbar if needed.

We also had to set the width of the box to 70vw.  This was
experimentally determined to give a reasonably wide area for the code
while not making the regular text wider than the screen.