Address #41:  Update displayed args for OPEN
In the HTML manual, the args for `OPEN` didn't show all of the
possible keyword args.  This was caused by the incorrect use of the
`@morekeys` and `@yetmorekeys` macros.  These macros take arguments,
and we forgot to put braces around the keywords that were meant as
arguments for the macros.