Fix #43:  Change WORDS_IN_PAGE to get bottom navbar more often
Reduce WORD_IN_PAGE to a smaller value so that the bottom navbar shows
up more often on short pages.  This is nice to have on mobile devices
with small screens.

Value determined experimentally by viewing short pages.  If the page
doesn't fit on the screen, the bottom navbar should be there.  This
includes portrait and landscape modes.

Did not try every possible short page.  As we find more, we can reduce
this value as needed.