Clean up issue with @defvrx in @deffn blocks
The original Scribe appeared to allow things like
```
@defun function
@defvrx {Hemlock Variable} var
<description>
@end defun
```

We translated this to
```
@deffn {Function} function
@deffnx {Hemlock Variable} var
<description>
@end deffn
```
While this looked right, the variable index didn't include `var`.
Instead, `var` was in the function index (as expected).

To work around this problem we do
```
@defvr {Hemlock Variable} var
@end defvr
@defun {Function} function
<description>
@end defun
```

Then everything shows up in the correct index.  There is an empty line
between the `@defvr` and `@defun`, but that seems unavoidable without
having to rewrite the description completely.