Address #47: More xrefs for with-compilation-unit
The macro `@macref` previously just used `@code`.  Change it to be
`@ref` to generate a cross-reference instead.  Replace many uses of
`@code{with-compilation-unit}` with `@macref`.  To make this all work,
we need to add an anchor for `with-compilation-unit`.

Also add a new macro `@keyref` to create a cross reference to a
keyword that prints the text as a keyword.  Use this new macro to
create references for `:context-declarations`.

Finally, the cross-references for context-declaration would take you
to the line following the section.  Move the anchor above the section
so that you can see the section heading when navigating to the link.