Fix #48: Update file time stamps for emacs
Update all the time stamps in the file to be automatically updated via
emacs.  This entails adding a file-local variable specification for
`time-stamp-pattern` in each file.  This also means adding `(add-hook
'write-file-hooks 'time-stamp) ` to your emacs init file to enable
updating the time stamp when the file is written.

We didn't update the individual news files because they haven't really
been updated and won't be since these are old news that shouldn't be
updated.