Fix #50: Fix up links to the original docs
The links for the original cmucl design pdf was a relative path which
doesn't exist on the new cmucl site.  Make the links point to
locations on cmucl.org so that they work on cmucl.common-lisp.dev.

These files still exist on
https://common-lisp.net/project/cmucl/downloads/doc/, but I don't know
how to add such things to cmucl.common-lisp.dev.  So I'm taking the
easy way and pointing to cmucl.org which syncs with this directory
anyway.