Update favicon
Generated a new favicon that looks a bit better.  We still use a
lambda character with a green background and black foreground.  The
green is the same as the green used in the header lines for pages.

This icon was generated via https://favicon.io/favicon-generator/
using the defaults, except for changing the font to Noto Seric, and
adjusting the font color and background, and a lambda character for
the text.