Fix #67:  Document xoroshiro128** generator
Add documentation for the xoroshiro128** generator and make a note
that this supersedes the xoroshiro128+ generator that is no longer
supported.

Also, since we build with texinfo 7.1 or greater, MathJax is supported
so update the Makefile so that MathJax is loaded to display nice HTML
math equations.  (We use this in the documentation for
random-state-jump.)