Address #68:  Add syntax highlighting of examples to docs
We add syntax highlighting of examples to cmu-user.

Syntax highlighting is done using highlight js and is set in
docs/texi2html.init.  This supports highlighting things in @lisp and
@example. This works for Lisp.  I think it should work for C as well,
but currently doesn't.  Also some @examples that contain mostly Lisp
don't get highlighting.  I'm not sure why.

Also, we add some CSS to add a background color to examples to set up
the example code from the surrounding text.  This is modeled after the
emacs manual.