Address #68:  Add highlighting of examples to hemlock
We add syntax highlighting of examples to the Hemlock user manual.

Syntax highlighting is done using highlight js and is set in
docs/texi2html.init.  This supports highlighting things in @lisp and
@example.

Also, we add some CSS to add a background color to examples to set up
the example code from the surrounding text.  This is modeled after the
emacs manual.